move Predicate and Variable into their own files
