+++ languages/python/oso/__init__.py
@@ -1,4 +1,3 @@
-from polar import polar_class
+from polar import polar_class, Variable, Predicate
 from .oso import Oso, OsoException
 from .extras import Http, PathMapper
-from polar.ffi import Variable, Predicate
