+++ languages/python/polar/__init__.py
@@ -1,6 +1,7 @@
 import json
 from pathlib import Path
 
-from .ffi import Predicate, Variable
 from .polar import Polar, polar_class
 from .query import Query, QueryResult
+from .variable import Variable
+from .predicate import Predicate
