+++ languages/python/polar/variable.py
@@ -1,0 +1,4 @@
+class Variable(str):
+    """An unbound variable type, can be used to query the KB for information"""
+
+    pass
